Script to bootstrap rethinkdb database and load CFA brigade data from API
@samba didn't see you had done the same db bootstrap until too late! Here I did the same think with the databases and tables named differently.

The import goes directly the CFA API and import all projects and organizaiton data. Right now this is about 3.4K rows of data, mostly projects.  The script to call is drop_and_reload_brigade_data.py to totally tear down and reload the db.

TODO: Still have to build out an API layer for the website to act as a go-between so RethinkDB can exist outside of localhost.