Use promises instead of timeout to wait for brigade details to resolve
Replace timeout call that waited 10s before caching $.get brigade details from teh Brigade Pulse API. Instead, store each resulting promis in an array, and then use $.when (with an apply), to wait for all objects in the array to resolve before continuing. This forces the caching to resolve programatically and predictably. It fixes an issue where, on slower connections, not all brigade data was cached to firebase, because the cache was called before all the $.get requests completed.