Fixed #69 Switching brigade.total_projects from String to Integer
This, coupled with the previous commit guaranteeing brigade results were cached, fixed iss #69.  By switching the types, it makes sure that when brigade.total_projects is used in an angular orderBy statement, it orders by the numeric order, not the lexical value. With lexical, the list ["7", "79", "80", "8"] would be ordered (descending) ["8", "80", "7", "79"].