Updated brigade details page with graphs and info a la @zachmccormick using static data
Created additional graphs outlining meetup attendance, github projects, and other information. These graphs are the same for every brigade, as they are loaded from static JSON objects in the brigades.controller.js.

Also added additional header information about the brigades, and changed the detail formatting to mirror the mockup @zachmccormick grabbed in the inline.

Additional fix - updated Contributors in the Projects to not display a trailing ",".